Add --[no-]deactivate-authorizations.
Resolves #5116.

Upon creation of an order (ACME v2) or authorization (ACME v1),
checks whether any authorization arrived with a status of `valid`
already. If found, these authorizations are deactivated and the
order is re-created.

Before an order is finalized (or otherwise fails), checks whether
any of the authorizations have a status of `valid` and deactivates
them before returning control to the user.

Automatically enabled by --staging and --dry-run.