Fix issue #7165 in _create_challenge_dirs()
Some operating systems won't reliably raise EEXIST or EISDIR if a directory exists when filesystem.mkdir() is called, instead raising `Read-only file system` errors or similar even when path is in fact writable. To work around this we include a check to see if directory already exists. See Issue #7165 for more information.